Fixes static IPs for Arduino Nano Every
The example program WebClient does not work on the Arduino Nano Every when
DHCP is not available or commented out - the versions of Ethernet::begin()
using a static IP address fail.  I tested the example on Uno, Mega2560 and
Micro, it works on all of them using either DHCP or static IP.

This modifies the begin() method so that it uses IPAddress::raw_address()
instead of platform conditional code in the same way the DHCP version of
begin() does. With this change static addresses work on the Nano Every too.